Fixed interface

+++ src/Lykke.Pkg.AzureRepositories/MerchantPayRequestRepository.cs
@@ -43,7 +43,8 @@ namespace Lykke.AzureRepositories
         public string SourceAddress { get; set; }
         public string DestinationAddress { get; set; }
         public string AssetPair { get; set; }
-        public decimal Amount { get; set; }
+
+        public double Amount { get; set; }
         public string AssetId { get; set; }
         public string SuccessUrl { get; set; }
         public string ErrorUrl { get; set; }
